+++ src/pages/components/PageTypePickerDialog/index.ts
@@ -1,0 +1,2 @@
+export * from "./PageTypePickerDialog";
+export { default } from "./PageTypePickerDialog";